# crates/dark-wallet-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dark-wallet-rest",
    "description": "REST wallet daemon for dark, an Ark protocol server for Bitcoin L2.\n\nExposes a curated subset of the gRPC surface as HTTP + JSON, with SSE for server-streamed events. Signer-path RPCs (tree nonces, tree signatures, forfeit signing) remain gRPC-only by design.\n\nAll endpoints return JSON. Binary fields (PSBTs, pubkeys, signatures) are hex-encoded strings unless noted. Amounts are denominated in satoshis.",
    "contact": {
      "name": "Lobby",
      "email": "lobby.clawy@gmail.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/txs": {
      "post": {
        "tags": [
          "txs"
        ],
        "summary": "Submit a signed Ark virtual tx",
        "operationId": "submit_tx",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTxRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitTxResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/txs/query": {
      "post": {
        "tags": [
          "txs"
        ],
        "summary": "List pending Ark txs matching an intent filter",
        "operationId": "query_pending_txs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentFilterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Matching pending txs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingTxsResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/txs/{id}": {
      "get": {
        "tags": [
          "txs"
        ],
        "summary": "Look up a pending Ark tx by id (best-effort)",
        "description": "Calls `ArkService.GetPendingTx` with an empty Intent filter and returns the first pending tx whose `ark_txid` matches. For authenticated queries use `POST /v1/txs/query`.",
        "operationId": "get_tx",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ark txid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending tx status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingTxResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not pending",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/txs/{id}/finalize": {
      "post": {
        "tags": [
          "txs"
        ],
        "operationId": "finalize_tx",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ark txid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeTxRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Finalized"
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FinalizeTxRequestDto": {
        "type": "object",
        "properties": {
          "final_checkpoint_txs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "IntentFilterDto": {
        "type": "object",
        "required": [
          "proof",
          "message"
        ],
        "properties": {
          "delegate_pubkey": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "proof": {
            "type": "string"
          }
        }
      },
...
      "PendingTxResponseDto": {
        "type": "object",
        "required": [
          "ark_txid",
          "status"
        ],
        "properties": {
          "ark_txid": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "PendingTxsResponse": {
        "type": "object",
        "required": [
          "pending_txs"
        ],
        "properties": {
          "pending_txs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PendingTxResponseDto"
            }
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 problem+json body.",
        "required": [
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code for this occurrence of the problem.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type."
          }
        }
      },
...
      "SubmitTxRequestDto": {
        "type": "object",
        "required": [
          "signed_ark_tx"
        ],
        "properties": {
          "checkpoint_txs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Checkpoint transactions in hex (optional)."
          },
          "signed_ark_tx": {
            "type": "string",
            "description": "Signed Ark virtual tx, hex encoded."
          }
        }
      },
      "SubmitTxResponseDto": {
        "type": "object",
        "required": [
          "ark_txid"
        ],
        "properties": {
          "ark_txid": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "macaroon"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}